+++ changelog.md
@@ -11,6 +11,7 @@ Codion Change Log
 - EntityPanel.addKeyEvent() and removeKeyEvent() added, editControlPanel() accessor removed.
 - EntityEditPanel, EntityTablePanel, EntityDialogs CRUD operations now performed in a background thread.
 - TabbedPanelLayout refactored.
+- TabbedPanelLayout.Builder.includeDetailTabPane() renamed includeDetailTabbedPane().
 ### is.codion.swing.framework.server.monitor
 - ServerMonitor and ServerMonitorPanel, tables migrated from plain JTable to FilteredTable.